tf2-api-python/tf2-api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pymysql==1.1.1
cryptography==42.0.7
